fix: Collection 추상 클래스 > 일반 클래스로 변경
- add(), addAll() 메소드를 Collection 클래스에서 구현
