+++ src/main/java/calculator/model/collection/Collection.java
@@ -1,0 +1,16 @@
+package calculator.model.collection;
+
+import java.util.List;
+
+public class Collection<T> {
+
+    protected List<T> values;
+
+    public void add(T t) {
+        this.values.add(t);
+    }
+
+    public void addAll(Collection<T> collection) {
+        this.values.addAll(collection.values);
+    }
+}
