+++ src/main/java/calculator/model/collection/Numbers.java
@@ -1,4 +1,6 @@
-package calculator.model;
+package calculator.model.collection;
+
+import calculator.model.Number;
 
 import java.util.ArrayList;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -7,16 +9,6 @@ public class Numbers extends Collection<Number> {
 
     public Numbers() {
         this.values = new ArrayList<>();
-    }
-
-    @Override
-    public void add(Number number) {
-        this.values.add(number);
-    }
-
-    @Override
-    public void addAll(Collection<Number> numbers) {
-        this.values.addAll(numbers.values);
     }
 
     public Number calculate(Symbols symbols) {
