+++ src/main/java/calculator/model/collection/Symbols.java
@@ -1,0 +1,16 @@
+package calculator.model.collection;
+
+import calculator.model.Symbol;
+
+import java.util.ArrayList;
+
+public class Symbols extends Collection<Symbol> {
+
+    public Symbols() {
+        this.values = new ArrayList<>();
+    }
+
+    public Symbol get(int index) {
+        return this.values.get(index);
+    }
+}
